Extract editing part of restricted editing mode plugin.
