+++ tests/restrictedediting.js
@@ -9,14 +9,7 @@ import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
 import ClassicTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/classictesteditor';
 
 import RestrictedEditing from './../src/restrictedediting';
-import { getData as getModelData, setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
-import { getData as getViewData } from '@ckeditor/ckeditor5-engine/src/dev-utils/view';
-import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
-import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
-import Typing from '@ckeditor/ckeditor5-typing/src/typing';
-import { assertEqualMarkup } from '@ckeditor/ckeditor5-utils/tests/_utils/utils';
-import UndoEditing from '@ckeditor/ckeditor5-undo/src/undoediting';
-import Command from '@ckeditor/ckeditor5-core/src/command';
+import RestrictedEditingEditing from './../src/restrictededitingediting';
 
 describe( 'RestrictedEditing', () => {
 	let editor, element;
@@ -44,292 +37,9 @@ describe( 'RestrictedEditing', () => {
 		it( 'should be loaded', () => {
 			expect( editor.plugins.get( RestrictedEditing ) ).to.be.instanceOf( RestrictedEditing );
 		} );
-	} );
 
-	describe( 'conversion', () => {
-		let model;
-
-		beforeEach( async () => {
-			editor = await VirtualTestEditor.create( { plugins: [ Paragraph, RestrictedEditing ] } );
-			model = editor.model;
-		} );
-
-		afterEach( () => {
-			return editor.destroy();
-		} );
-
-		describe( 'upcast', () => {
-			it( 'should convert <span class="ck-restricted-editing-exception"> to marker', () => {
-				editor.setData( '<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>' );
-
-				expect( model.markers.has( 'restricted-editing-exception:1' ) ).to.be.true;
-
-				const marker = model.markers.get( 'restricted-editing-exception:1' );
-
-				expect( marker.getStart().path ).to.deep.equal( [ 0, 4 ] );
-				expect( marker.getEnd().path ).to.deep.equal( [ 0, 7 ] );
-			} );
-
-			it( 'should convert multiple <span class="ck-restricted-editing-exception">', () => {
-				editor.setData(
-					'<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>' +
-					'<p>ABCDEF<span class="ck-restricted-editing-exception">GHIJK</span>LMNOPQRST</p>'
-				);
-
-				expect( model.markers.has( 'restricted-editing-exception:1' ) ).to.be.true;
-				expect( model.markers.has( 'restricted-editing-exception:2' ) ).to.be.true;
-
-				// Data for the first marker is the same as in previous tests so no need to test it again.
-				const secondMarker = model.markers.get( 'restricted-editing-exception:2' );
-
-				expect( secondMarker.getStart().path ).to.deep.equal( [ 1, 6 ] );
-				expect( secondMarker.getEnd().path ).to.deep.equal( [ 1, 11 ] );
-			} );
-
-			it( 'should not convert other <span> elements', () => {
-				editor.setData( '<p>foo <span class="foo bar">bar</span> baz</p>' );
-
-				expect( model.markers.has( 'restricted-editing-exception:1' ) ).to.be.false;
-			} );
-		} );
-
-		describe( 'downcast', () => {
-			it( 'should convert model marker to <span>', () => {
-				setModelData( model, '<paragraph>foo bar baz</paragraph>' );
-
-				const paragraph = model.document.getRoot().getChild( 0 );
-
-				model.change( writer => {
-					writer.addMarker( 'restricted-editing-exception:1', {
-						range: writer.createRange( writer.createPositionAt( paragraph, 4 ), writer.createPositionAt( paragraph, 7 ) ),
-						usingOperation: true,
-						affectsData: true
-					} );
-				} );
-
-				const expectedView = '<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>';
-				expect( editor.getData() ).to.equal( expectedView );
-				expect( getViewData( editor.editing.view, { withoutSelection: true } ) ).to.equal( expectedView );
-			} );
-
-			it( 'converted <span> should be the outermost attribute element', () => {
-				editor.conversion.for( 'downcast' ).attributeToElement( { model: 'bold', view: 'b' } );
-				setModelData( model, '<paragraph><$text bold="true">foo bar baz</$text></paragraph>' );
-
-				const paragraph = model.document.getRoot().getChild( 0 );
-
-				model.change( writer => {
-					writer.addMarker( 'restricted-editing-exception:1', {
-						range: writer.createRange( writer.createPositionAt( paragraph, 0 ), writer.createPositionAt( paragraph, 'end' ) ),
-						usingOperation: true,
-						affectsData: true
-					} );
-				} );
-
-				const expectedView = '<p><span class="ck-restricted-editing-exception"><b>foo bar baz</b></span></p>';
-				expect( editor.getData() ).to.equal( expectedView );
-				expect( getViewData( editor.editing.view, { withoutSelection: true } ) ).to.equal( expectedView );
-			} );
+		it( 'should loaded RestrictedEditingEditing plugin', () => {
+			expect( editor.plugins.get( RestrictedEditingEditing ) ).to.be.instanceOf( RestrictedEditingEditing );
 		} );
 	} );
-
-	describe( 'editing behavior', () => {
-		let model;
-
-		beforeEach( async () => {
-			editor = await VirtualTestEditor.create( { plugins: [ Paragraph, Typing, UndoEditing, RestrictedEditing ] } );
-			model = editor.model;
-		} );
-
-		afterEach( async () => {
-			await editor.destroy();
-		} );
-
-		it( 'should keep markers in the view when editable region is edited', () => {
-			setModelData( model,
-				'<paragraph>foo bar baz</paragraph>' +
-				'<paragraph>xxx y[]yy zzz</paragraph>'
-			);
-
-			const firstParagraph = model.document.getRoot().getChild( 0 );
-			const secondParagraph = model.document.getRoot().getChild( 1 );
-
-			model.change( writer => {
-				writer.addMarker( 'restricted-editing-exception:1', {
-					range: writer.createRange( writer.createPositionAt( firstParagraph, 4 ), writer.createPositionAt( firstParagraph, 7 ) ),
-					usingOperation: true,
-					affectsData: true
-				} );
-				writer.addMarker( 'restricted-editing-exception:2', {
-					range: writer.createRange(
-						writer.createPositionAt( secondParagraph, 4 ),
-						writer.createPositionAt( secondParagraph, 7 )
-					),
-					usingOperation: true,
-					affectsData: true
-				} );
-			} );
-
-			model.change( writer => {
-				model.insertContent( writer.createText( 'R', model.document.selection.getAttributes() ) );
-			} );
-
-			const expectedView = '<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>' +
-				'<p>xxx <span class="ck-restricted-editing-exception">yRyy</span> zzz</p>';
-
-			expect( editor.getData() ).to.equal( expectedView );
-			expect( getViewData( editor.editing.view, { withoutSelection: true } ) ).to.equal( expectedView );
-		} );
-
-		it( 'should block user typing outside exception markers', () => {
-			setModelData( model, '<paragraph>foo []bar baz</paragraph>' );
-
-			editor.execute( 'input', { text: 'X' } );
-
-			assertEqualMarkup( getModelData( model ), '<paragraph>foo []bar baz</paragraph>' );
-		} );
-
-		it( 'should not block user typing inside exception marker', () => {
-			setModelData( model, '<paragraph>[]foo bar baz</paragraph>' );
-			const firstParagraph = model.document.getRoot().getChild( 0 );
-
-			model.change( writer => {
-				writer.addMarker( 'restricted-editing-exception:1', {
-					range: writer.createRange( writer.createPositionAt( firstParagraph, 4 ), writer.createPositionAt( firstParagraph, 7 ) ),
-					usingOperation: true,
-					affectsData: true
-				} );
-			} );
-
-			model.change( writer => {
-				writer.setSelection( firstParagraph, 5 );
-			} );
-			editor.execute( 'input', { text: 'X' } );
-
-			assertEqualMarkup( getModelData( model ), '<paragraph>foo bX[]ar baz</paragraph>' );
-		} );
-	} );
-
-	describe( 'commands integration', () => {
-		let model, firstParagraph;
-
-		beforeEach( async () => {
-			editor = await VirtualTestEditor.create( { plugins: [ Paragraph, Typing, UndoEditing, RestrictedEditing ] } );
-			model = editor.model;
-
-			setModelData( model, '<paragraph>[]foo bar baz</paragraph>' );
-			firstParagraph = model.document.getRoot().getChild( 0 );
-
-			model.change( writer => {
-				writer.addMarker( 'restricted-editing-exception:1', {
-					range: writer.createRange(
-						writer.createPositionAt( firstParagraph, 4 ),
-						writer.createPositionAt( firstParagraph, 7 ) ),
-					usingOperation: true,
-					affectsData: true
-				} );
-			} );
-		} );
-
-		afterEach( async () => {
-			await editor.destroy();
-		} );
-
-		describe( 'undo', () => {
-			it( 'should be enabled outside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 1 );
-				} );
-
-				expect( editor.commands.get( 'undo' ).isEnabled ).to.be.true;
-			} );
-
-			it( 'should be enabled inside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 5 );
-				} );
-
-				expect( editor.commands.get( 'undo' ).isEnabled ).to.be.true;
-			} );
-		} );
-
-		describe( 'redo', () => {
-			beforeEach( () => {
-				editor.commands.add( 'redo', buildFakeCommand( editor ) );
-			} );
-
-			it( 'should be enabled outside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 1 );
-				} );
-
-				expect( editor.commands.get( 'redo' ).isEnabled ).to.be.true;
-			} );
-
-			it( 'should be enabled inside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 5 );
-				} );
-
-				expect( editor.commands.get( 'redo' ).isEnabled ).to.be.true;
-			} );
-		} );
-
-		describe( 'bold', () => {
-			beforeEach( () => {
-				editor.commands.add( 'bold', buildFakeCommand( editor ) );
-			} );
-
-			it( 'should be disabled outside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 1 );
-				} );
-
-				expect( editor.commands.get( 'bold' ).isEnabled ).to.be.false;
-			} );
-
-			it( 'should be enabled inside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 5 );
-				} );
-
-				expect( editor.commands.get( 'bold' ).isEnabled ).to.be.true;
-			} );
-		} );
-
-		describe( 'other', () => {
-			beforeEach( () => {
-				editor.commands.add( 'other', buildFakeCommand( editor ) );
-			} );
-
-			it( 'should be disabled outside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 1 );
-				} );
-
-				expect( editor.commands.get( 'other' ).isEnabled ).to.be.false;
-			} );
-
-			it( 'should be disabled inside exception marker', () => {
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 1 );
-				} );
-				model.change( writer => {
-					writer.setSelection( firstParagraph, 5 );
-				} );
-
-				expect( editor.commands.get( 'other' ).isEnabled ).to.be.false;
-			} );
-		} );
-	} );
-
-	class FakeCommand extends Command {
-		refresh() {
-			this.isEnabled = true;
-		}
-	}
-
-	function buildFakeCommand( editor ) {
-		return new FakeCommand( editor );
-	}
 } );
